api/pyproject.toml
[project]
name = "aegis-health-api"
version = "0.1.0"
description = "FastAPI service for Aegis Health"
authors = [{ name = "Aegis Health" }]
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.27.0",
    "pydantic==2.4.2",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "httpx[http2]>=0.24.1",
    "fastapi-socketio>=0.0.10",
    "aio-pika>=9.2.0",
    "opentelemetry-api>=1.18.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-exporter-otlp>=1.18.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
